src directory fully refactored.
