+++ src/pages/Companies/Companies.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react"
 import { Link, useNavigate } from "react-router-dom";
-import { Company } from "../../Interfaces";
+import { Company } from "../../constants/Interfaces";
 import { useUserLoggedContext } from "../../context/UserLoggedContext";
 import { fetchCompanies } from "../../constants/trackerApiCalls";
 import { useErrorContext } from "../../context/ErrorContext";
