+++ src/pages/Companies/components/NewCompany.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { fetchCompanies, createCompany } from "../../../constants/trackerApiCalls";
-import { CompanyAttributes } from "../../../Interfaces";
+import { CompanyAttributes } from "../../../constants/Interfaces";
 import { useUserLoggedContext } from "../../../context/UserLoggedContext";
 import { useErrorContext } from "../../../context/ErrorContext";
