+++ src/pages/Contacts/Contacts.tsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { useUserLoggedContext } from "../../context/UserLoggedContext";
-import { ContactData, UserInformationProps } from "../../Interfaces";
-import { fetchContacts } from '../../apiCalls';
+import { ContactData, UserInformationProps } from "../../constants/Interfaces";
+import { fetchContacts } from '../../constants/apiCalls';
 
 function Contacts({ userData }: UserInformationProps) {
   const [contacts, setContacts] = useState<ContactData[] | []>([]);
