+++ src/pages/Contacts/components/NewContact.tsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 import { useUserLoggedContext } from "../../../context/UserLoggedContext";
-import { fetchCompanies, fetchNewContact } from "../../../apiCalls";
-import { UserInformationProps, FormInputData } from "../../../Interfaces";
+import { fetchCompanies, fetchNewContact } from "../../../constants/apiCalls";
+import { UserInformationProps, FormInputData } from "../../../constants/Interfaces";
 import CompanyModal from "../../Companies/modals/CompanyModal";
 
 const NewContact = ({ userData }: UserInformationProps) => {
